# openapi.json
{"openapi":"3.0.1","info":{"title":"iBank API","description":"Documentation iBank API v1.0","version":"1.0"},"servers":[{"url":"http://localhost:8080/ibank","description":"Generated server url"}],"tags":[{"name":"customers","description":"Customer API operations"},{"name":"accounts","description":"Account API operations"}],"paths":{"/api/customer/create":{"post":{"tags":["customers"],"summary":"Create a new customer","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"201":{"description":"Customer created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}},"400":{"description":"Invalid customer supplied","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/api/account/transfer":{"post":{"tags":["accounts"],"summary":"Transfer amount between accounts","operationId":"transfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequest"}}},"required":true},"responses":{"400":{"description":"Invalid account(s) supplied","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"200":{"description":"All deposits retrieved","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}},"404":{"description":"Account(s) not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Transaction"}}}}}}},"/api/account/create":{"post":{"tags":["accounts"],"summary":"Create a new account","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}},"required":true},"responses":{"201":{"description":"Account created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}},"400":{"description":"Invalid account supplied","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/customer/{customerId}/accounts":{"get":{"tags":["customers"],"summary":"Retrieve customer accounts","operationId":"accounts","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Invalid customer id","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}},"404":{"description":"Customer not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}},"200":{"description":"All customer accounts retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}}},"/api/customer/all":{"get":{"tags":["customers"],"summary":"Retrieve all customers","operationId":"all","responses":{"200":{"description":"All customers retrieved","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}},"500":{"description":"Internal error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Customer"}}}}}}}},"/api/account/{accountNumber}":{"get":{"tags":["accounts"],"summary":"Retrieve account by number","operationId":"findByNumber","parameters":[{"name":"accountNumber","in":"path","description":"account number to be searched","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All accounts retrieved","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Invalid account number","content":{"application/json":{"schema":{"type":"string"}}}},"404":{"description":"Account not found","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/account/{accountNumber}/withdrawals":{"get":{"tags":["accounts"],"summary":"Retrieve withdrawals by account number","operationId":"withdrawals","parameters":[{"name":"accountNumber","in":"path","description":"account number to be searched","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All withdrawals retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"400":{"description":"Invalid account number","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"404":{"description":"Account not found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/account/{accountNumber}/transactions":{"get":{"tags":["accounts"],"summary":"Retrieve transactions by account number","operationId":"transactions","parameters":[{"name":"accountNumber","in":"path","description":"account number to be searched","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All transactions retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"400":{"description":"Invalid account number","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"404":{"description":"Account not found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/account/{accountNumber}/deposits":{"get":{"tags":["accounts"],"summary":"Retrieve deposits by account number","operationId":"deposits","parameters":[{"name":"accountNumber","in":"path","description":"account number to be searched","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All deposits retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"400":{"description":"Invalid account number","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}},"404":{"description":"Account not found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}},"/api/account/all":{"get":{"tags":["accounts"],"summary":"Retrieve all accounts","operationId":"all_1","responses":{"200":{"description":"All accounts retrieved","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}},"500":{"description":"Internal error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Account"}}}}}}}},"/api/customer/{customerId}":{"delete":{"tags":["customers"],"summary":"Delete customer by id","operationId":"delete","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Invalid customer id"},"404":{"description":"Customer not found"},"200":{"description":"Customer deleted"}}}}},"components":{"schemas":{"Customer":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"TransactionRequest":{"required":["amount","fromAccount","toAccount"],"type":"object","properties":{"fromAccount":{"type":"string"},"toAccount":{"type":"string"},"amount":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"}}},"Transaction":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"number"},"time":{"type":"string","format":"date-time"},"fromAccount":{"type":"string"},"toAccount":{"type":"string"}}},"Account":{"required":["customer"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"currentBalance":{"minimum":0.0,"exclusiveMinimum":false,"type":"number"},"customer":{"$ref":"#/components/schemas/Customer"}}}}}}